refactor(modules): seperate modules into components pages and utils folders
